pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "special-agent-ops"
version = "0.1.0"
description = "kubectl-like CLI for agent manifests, codegen, deploy, and gateway"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "rich>=13.0",
    "pydantic-ai>=0.0.24",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "redis>=5.2",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "testcontainers[redis]>=4.8",
]

[project.scripts]
agentctl = "agentctl.cli.app:main"
agent-gateway = "agent_gateway.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentctl", "src/agent_gateway"]

[tool.hatch.build.targets.sdist]
include = ["src", "tests", "examples", "README.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
